[기능추가]: 파일과 내용을 입력받아 해당 파일에 내용을 검색하는 기능 추가하고, 기존 read_file 메소드를 FileControl.py로 이동하고 Exception을 rasie하도록 변경했습니다.
